pixtexed: start toolbar

+++ pixtexed/color.h
@@ -24,6 +24,6 @@ uint32_t color32_g(uint32_t c);
 uint32_t color32_b(uint32_t c);
 uint32_t color32_a(uint32_t c);
 
-inline uint32_t color32(uint8_t r, uint8_t g, uint8_t b, uint8_t a) { return (r)|(g<<8)|(b<<16)|(a<<24); }
+inline uint32_t color32(uint8_t r, uint8_t g, uint8_t b, uint8_t a) { return (uint32_t)(r)|(uint32_t)(g<<8)|(uint32_t)(b<<16)|(uint32_t)(a<<24); }
 
 #endif
